Create a `fs` utils module
