+++ collector/src/execute.rs
@@ -3,11 +3,12 @@
 use crate::{Compiler, Profile, Scenario};
 use anyhow::{bail, Context};
 use collector::category::Category;
-use collector::command_output;
 use collector::etw_parser;
+use collector::{command_output, utils};
 use database::{PatchName, QueryLabel};
 use futures::stream::FuturesUnordered;
 use futures::stream::StreamExt;
+use std::cmp;
 use std::collections::HashMap;
 use std::env;
 use std::fmt;
@@ -19,93 +20,10 @@ use std::path::{Path, PathBuf};
 use std::process::{self, Command};
 use std::str;
 use std::time::Duration;
-use std::{cmp, ffi::OsStr, path::Component};
 use tempfile::TempDir;
 use tokio::runtime::Runtime;
 
 mod rustc;
-
-#[cfg(windows)]
-pub fn rename<P: AsRef<Path>, Q: AsRef<Path>>(from: P, to: Q) -> anyhow::Result<()> {
-    let (from, to) = (from.as_ref(), to.as_ref());
-
-    let ctx = format!("renaming file {:?} to {:?}", from, to);
-
-    if fs::metadata(from)?.is_file() {
-        return Ok(fs::rename(from, to).with_context(|| ctx.clone())?);
-    }
-
-    collector::robocopy(from, to, &[&"/move"]).with_context(|| ctx.clone())
-}
-
-#[cfg(unix)]
-pub fn rename<P: AsRef<Path>, Q: AsRef<Path>>(from: P, to: Q) -> anyhow::Result<()> {
-    let (from, to) = (from.as_ref(), to.as_ref());
-    if fs::rename(from, to).is_err() {
-        // This is necessary if from and to are on different
-        // mount points (e.g., if /tmp is in tmpfs instead of on
-        // the same disk). We don't want to implement a full recursive solution
-        // to copying directories, so just shell out to `mv`.
-        let ctx = format!("mv {:?} {:?}", from, to);
-        let status = Command::new("mv")
-            .arg(from)
-            .arg(to)
-            .status()
-            .with_context(|| ctx.clone())?;
-        if !status.success() {
-            anyhow::bail!("mv {:?} {:?}: {:?}", from, to, status);
-        }
-    }
-
-    Ok(())
-}
-
-fn touch(path: &Path) -> anyhow::Result<()> {
-    log::trace!("touching file {:?}", path);
-
-    filetime::set_file_mtime(path, filetime::FileTime::now())
-        .with_context(|| format!("touching file {:?}", path))?;
-
-    Ok(())
-}
-
-fn touch_all(path: &Path) -> anyhow::Result<()> {
-    fn is_valid(path: &Path) -> bool {
-        let target_dir = Component::Normal(OsStr::new("target"));
-
-        // Don't touch files in `target/`, since they're likely generated by build scripts and might be from a dependency.
-        if path.components().any(|component| component == target_dir) {
-            return false;
-        }
-
-        if let Some(extn) = path.extension() {
-            if extn.to_str() == Some("rs") {
-                // Don't touch build scripts, which confuses the wrapped rustc.
-                return path.file_name() != Some(OsStr::new("build.rs"));
-            }
-        }
-
-        false
-    }
-
-    for entry in walkdir::WalkDir::new(path) {
-        let entry = entry?;
-        let path = entry.path();
-
-        // We also delete the cmake caches to avoid errors when moving directories around.
-        // This might be a bit slower but at least things build
-        if path.file_name() == Some(OsStr::new("CMakeCache.txt")) {
-            fs::remove_file(path)
-                .with_context(|| format!("deleting cmake caches in {:?}", path))?;
-        }
-
-        if is_valid(path) {
-            touch(path)?;
-        }
-    }
-
-    Ok(())
-}
 
 fn default_runs() -> usize {
     3
@@ -444,9 +362,9 @@ impl<'a> CargoProcess<'a> {
                 // in-tree (e.g., in the case of the servo crates there are a lot of
                 // other components).
                 if let Some(file) = &self.touch_file {
-                    touch(&self.cwd.join(Path::new(&file)))?;
+                    utils::fs::touch(&self.cwd.join(Path::new(&file)))?;
                 } else {
-                    touch_all(
+                    utils::fs::touch_all(
                         &self.cwd.join(
                             Path::new(&self.manifest_path)
                                 .parent()
@@ -901,7 +819,7 @@ impl<'a> Processor for BenchProcessor<'a> {
 }
 
 fn store_documentation_size_into_stats(stats: &mut Stats, doc_dir: &Path) {
-    match get_file_count_and_size(doc_dir) {
+    match utils::fs::get_file_count_and_size(doc_dir) {
         Ok((count, size)) => {
             stats.insert("size:doc_files_count".to_string(), count as f64);
             stats.insert("size:doc_bytes".to_string(), size as f64);
@@ -912,30 +830,6 @@ fn store_documentation_size_into_stats(stats: &mut Stats, doc_dir: &Path) {
             error
         ),
     }
-}
-
-/// Counts the number of files and the total size of all files within the given `path`.
-/// File size is counted as the actual size in bytes, i.e. the size returned by
-/// [std::path::Path::metadata].
-///
-/// Returns (file_count, size).
-pub fn get_file_count_and_size(path: &Path) -> std::io::Result<(u64, u64)> {
-    let (count, size) = if path.is_dir() {
-        let mut file_count = 0;
-        let mut total_size = 0;
-        for entry in fs::read_dir(&path)? {
-            let path = entry?.path();
-            let (count, size) = get_file_count_and_size(&path)?;
-            file_count += count;
-            total_size += size;
-        }
-        (file_count, total_size)
-    } else if path.is_file() {
-        (1, path.metadata()?.len())
-    } else {
-        (0, 0)
-    };
-    Ok((count, size))
 }
 
 fn store_artifact_sizes_into_stats(stats: &mut Stats, profile: &SelfProfile) {
@@ -1010,7 +904,7 @@ impl<'a> Processor for ProfileProcessor<'a> {
                     fs::remove_dir_all(&zsp_dir)?;
                 }
 
-                rename(&tmp_zsp_dir, &zsp_dir)?;
+                utils::fs::rename(&tmp_zsp_dir, &zsp_dir)?;
 
                 // Rename the data files. There should be exactly three.
                 let mut num_files = 0;
@@ -1020,13 +914,13 @@ impl<'a> Processor for ProfileProcessor<'a> {
                     let filename_str = filename.to_str().unwrap();
                     let path = filepath(&zsp_dir, filename_str);
                     if filename_str.ends_with(".events") {
-                        rename(path, filepath(&zsp_dir, "Zsp.events"))?;
+                        utils::fs::rename(path, filepath(&zsp_dir, "Zsp.events"))?;
                     } else if filename_str.ends_with(".string_data") {
-                        rename(path, filepath(&zsp_dir, "Zsp.string_data"))?;
+                        utils::fs::rename(path, filepath(&zsp_dir, "Zsp.string_data"))?;
                     } else if filename_str.ends_with(".string_index") {
-                        rename(path, filepath(&zsp_dir, "Zsp.string_index"))?;
+                        utils::fs::rename(path, filepath(&zsp_dir, "Zsp.string_index"))?;
                     } else if filename_str.ends_with(".mm_profdata") {
-                        rename(path, filepath(&zsp_dir, "Zsp.mm_profdata"))?;
+                        utils::fs::rename(path, filepath(&zsp_dir, "Zsp.mm_profdata"))?;
                     } else {
                         panic!("unexpected file {:?}", path);
                     }
@@ -1045,13 +939,13 @@ impl<'a> Processor for ProfileProcessor<'a> {
                 let mut flamegraph_cmd = Command::new("flamegraph");
                 flamegraph_cmd.arg(&zsp_files_prefix);
                 flamegraph_cmd.status().context("flamegraph")?;
-                rename("rustc.svg", flamegraph_file)?;
+                utils::fs::rename("rustc.svg", flamegraph_file)?;
 
                 // Run `crox`.
                 let mut crox_cmd = Command::new("crox");
                 crox_cmd.arg(&zsp_files_prefix);
                 crox_cmd.status().context("crox")?;
-                rename("chrome_profiler.json", crox_file)?;
+                utils::fs::rename("chrome_profiler.json", crox_file)?;
             }
 
             // `-Ztime-passes` output is redirected (via rustc-fake) to a file
@@ -1087,7 +981,7 @@ impl<'a> Processor for ProfileProcessor<'a> {
                 if opout_dir.exists() {
                     fs::remove_dir_all(&opout_dir)?;
                 }
-                rename(&tmp_opout_dir, &opout_dir)?;
+                utils::fs::rename(&tmp_opout_dir, &opout_dir)?;
 
                 let mut session_dir_arg = "--session-dir=".to_string();
                 session_dir_arg.push_str(opout_dir.to_str().unwrap());
@@ -1334,7 +1228,7 @@ impl Benchmark {
 
     #[cfg(windows)]
     fn copy(from: &Path, to: &Path) -> anyhow::Result<()> {
-        collector::robocopy(from, to, &[])
+        crate::utils::fs::robocopy(from, to, &[])
     }
 
     #[cfg(unix)]
@@ -1801,32 +1695,4 @@ pub struct QueryData {
     pub invocation_count: u32,
     pub blocked_time: Duration,
     pub incremental_load_time: Duration,
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::execute::get_file_count_and_size;
-    use std::path::PathBuf;
-
-    #[test]
-    fn test_get_file_count_and_size() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let root = dir.path();
-
-        let write = |path: PathBuf, size: usize| {
-            std::fs::create_dir_all(path.parent().unwrap()).unwrap();
-            std::fs::write(path, vec![0u8; size].as_slice()).unwrap();
-        };
-
-        write(root.join("a/b/c.rs"), 1024);
-        write(root.join("a/b/d.rs"), 16);
-        write(root.join("a/x.rs"), 32);
-        write(root.join("b/x.rs"), 64);
-        write(root.join("b/x2.rs"), 64);
-        write(root.join("x.rs"), 128);
-
-        let (files, size) = get_file_count_and_size(root).unwrap();
-        assert_eq!(files, 6);
-        assert_eq!(size, 1024 + 16 + 32 + 64 + 64 + 128);
-    }
 }
