+++ collector/src/lib.rs
@@ -8,7 +8,7 @@ use std::process::{self, Command};
 pub mod api;
 pub mod category;
 pub mod etw_parser;
-mod utils;
+pub mod utils;
 
 use process::Stdio;
 
@@ -153,34 +153,6 @@ pub fn run_command(cmd: &mut Command) -> anyhow::Result<()> {
     if !status.success() {
         return Err(anyhow::anyhow!("expected success {:?}", status));
     }
-    Ok(())
-}
-
-#[cfg(windows)]
-pub fn robocopy(
-    from: &std::path::Path,
-    to: &std::path::Path,
-    extra_args: &[&dyn AsRef<std::ffi::OsStr>],
-) -> anyhow::Result<()> {
-    let mut cmd = Command::new("robocopy");
-    cmd.arg(from).arg(to).arg("/s").arg("/e");
-
-    for arg in extra_args {
-        cmd.arg(arg.as_ref());
-    }
-
-    let output = run_command_with_output(&mut cmd)?;
-
-    if output.status.code() >= Some(8) {
-        // robocopy returns 0-7 on success
-        return Err(anyhow::anyhow!(
-            "expected success, got {}\n\nstderr={}\n\n stdout={}",
-            output.status,
-            String::from_utf8_lossy(&output.stderr),
-            String::from_utf8_lossy(&output.stdout)
-        ));
-    }
-
     Ok(())
 }
