+++ collector/src/main.rs
@@ -4,6 +4,7 @@ use anyhow::{bail, Context};
 use clap::Parser;
 use collector::api::next_artifact::NextArtifact;
 use collector::category::Category;
+use collector::utils;
 use database::{ArtifactId, Commit, CommitType, Pool};
 use log::debug;
 use rayon::iter::{IndexedParallelIterator, IntoParallelRefIterator, ParallelIterator};
@@ -1335,7 +1336,7 @@ fn download_from_git(target: &Path, url: &str) -> anyhow::Result<()> {
         log::error!("Could not delete .git directory: {error:?}");
     }
 
-    execute::rename(&tmpdir, &target)?;
+    utils::fs::rename(&tmpdir, &target)?;
     Ok(())
 }
 
@@ -1365,7 +1366,7 @@ fn download_from_crates_io(target_dir: &Path, krate: &str, version: &str) -> any
     // under <crate-name>-<version> directory.
     let unpacked_dir = tmpdir.path().join(format!("{krate}-{version}"));
     generate_lockfile(&unpacked_dir);
-    execute::rename(&unpacked_dir, &target_dir)?;
+    utils::fs::rename(&unpacked_dir, &target_dir)?;
 
     Ok(())
 }
