+++ collector/src/utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod fs;
 pub mod read2;
